refactor(core): remove window endpoints
