+++ .changes/ipc-scope-remove-enable-tauri-api.md
@@ -1,0 +1,6 @@
+---
+"tauri": patch
+"tauri-utils": patch
+---
+
+Remove `enable_tauri_api` from the IPC scope.
