+++ .changes/remove-macros-command-module.md
@@ -1,0 +1,5 @@
+---
+"tauri-macros": patch
+---
+
+Removed the module command macros.
