+++ .changes/remove-window.md
@@ -1,0 +1,6 @@
+---
+"tauri": patch
+"api": patch
+---
+
+Moved the `window` JS APIs to its own plugin in the plugins-workspace repository.
