+++ core/tauri-macros/src/lib.rs
@@ -4,10 +4,9 @@
 
 use crate::context::ContextItems;
 use proc_macro::TokenStream;
-use syn::{parse_macro_input, DeriveInput, ItemFn};
+use syn::{parse_macro_input, DeriveInput};
 
 mod command;
-mod command_module;
 mod mobile;
 mod runtime;
 
@@ -80,41 +79,4 @@ pub fn default_runtime(attributes: TokenStream, input: TokenStream) -> TokenStre
   let attributes = parse_macro_input!(attributes as runtime::Attributes);
   let input = parse_macro_input!(input as DeriveInput);
   runtime::default_runtime(attributes, input).into()
-}
-
-/// Prepares the command module enum.
-#[doc(hidden)]
-#[proc_macro_derive(CommandModule, attributes(cmd))]
-pub fn derive_command_module(input: TokenStream) -> TokenStream {
-  let input = parse_macro_input!(input as DeriveInput);
-  command_module::generate_run_fn(input)
-}
-
-/// Adds a `run` method to an enum (one of the tauri endpoint modules).
-/// The `run` method takes a `tauri::endpoints::InvokeContext`
-/// and returns a `tauri::Result<tauri::endpoints::InvokeResponse>`.
-/// It matches on each enum variant and call a method with name equal to the variant name, lowercased and snake_cased,
-/// passing the context and the variant's fields as arguments.
-/// That function must also return the same `Result<InvokeResponse>`.
-#[doc(hidden)]
-#[proc_macro_attribute]
-pub fn command_enum(_: TokenStream, input: TokenStream) -> TokenStream {
-  let input = parse_macro_input!(input as DeriveInput);
-  command_module::generate_command_enum(input)
-}
-
-#[doc(hidden)]
-#[proc_macro_attribute]
-pub fn module_command_handler(attributes: TokenStream, input: TokenStream) -> TokenStream {
-  let attributes = parse_macro_input!(attributes as command_module::HandlerAttributes);
-  let input = parse_macro_input!(input as ItemFn);
-  command_module::command_handler(attributes, input).into()
-}
-
-#[doc(hidden)]
-#[proc_macro_attribute]
-pub fn module_command_test(attributes: TokenStream, input: TokenStream) -> TokenStream {
-  let attributes = parse_macro_input!(attributes as command_module::HandlerTestAttributes);
-  let input = parse_macro_input!(input as ItemFn);
-  command_module::command_test(attributes, input).into()
 }
