+++ examples/api/src/App.svelte
@@ -1,19 +1,11 @@
 <script>
   import { writable } from 'svelte/store'
-  import { appWindow, getCurrent } from '@tauri-apps/api/window'
 
   import Welcome from './views/Welcome.svelte'
-  import Cli from './views/Cli.svelte'
   import Communication from './views/Communication.svelte'
-  import Window from './views/Window.svelte'
   import WebRTC from './views/WebRTC.svelte'
 
   import { onMount } from 'svelte'
-  import { listen } from '@tauri-apps/api/event'
-
-  appWindow.onFileDropEvent((event) => {
-    onMessage(`File drop: ${JSON.stringify(event.payload)}`)
-  })
 
   const userAgent = navigator.userAgent.toLowerCase()
   const isMobile = userAgent.includes('android') || userAgent.includes('iphone')
@@ -29,16 +21,6 @@
       component: Communication,
       icon: 'i-codicon-radio-tower'
     },
-    !isMobile && {
-      label: 'CLI',
-      component: Cli,
-      icon: 'i-codicon-terminal'
-    },
-    !isMobile && {
-      label: 'Window',
-      component: Window,
-      icon: 'i-codicon-window'
-    },
     {
       label: 'WebRTC',
       component: WebRTC,
@@ -49,25 +31,6 @@
   let selected = views[0]
   function select(view) {
     selected = view
-  }
-
-  // Window controls
-  let isWindowMaximized
-  onMount(async () => {
-    const window = getCurrent()
-    isWindowMaximized = await window.isMaximized()
-    listen('tauri://resize', async () => {
-      isWindowMaximized = await window.isMaximized()
-    })
-  })
-
-  function minimize() {
-    getCurrent().minimize()
-  }
-
-  async function toggleMaximize() {
-    const window = getCurrent()
-    ;(await window.isMaximized()) ? window.unmaximize() : window.maximize()
   }
 
   // dark/light
@@ -141,8 +104,6 @@
     document.addEventListener('mousemove', moveHandler)
   }
 
-  const isWindows = navigator.appVersion.includes('Win')
-
   // mobile
   let isSideBarOpen = false
   let sidebar
@@ -211,52 +172,6 @@
   }
 </script>
 
-<!-- custom titlebar for Windows -->
-{#if isWindows}
-  <div
-    class="w-screen select-none h-8 pl-2 flex justify-between items-center absolute text-primaryText dark:text-darkPrimaryText"
-    data-tauri-drag-region
-  >
-    <span class="lt-sm:pl-10 text-darkPrimaryText">Tauri API Validation</span>
-    <span
-      class="
-      h-100%
-      children:h-100% children:w-12 children:inline-flex
-      children:items-center children:justify-center"
-    >
-      <span
-        title={isDark ? 'Switch to Light mode' : 'Switch to Dark mode'}
-        class="hover:bg-hoverOverlay active:bg-hoverOverlayDarker dark:hover:bg-darkHoverOverlay dark:active:bg-darkHoverOverlayDarker"
-        on:click={toggleDark}
-      >
-        {#if isDark}
-          <div class="i-ph-sun" />
-        {:else}
-          <div class="i-ph-moon" />
-        {/if}
-      </span>
-      <span
-        title="Minimize"
-        class="hover:bg-hoverOverlay active:bg-hoverOverlayDarker dark:hover:bg-darkHoverOverlay dark:active:bg-darkHoverOverlayDarker"
-        on:click={minimize}
-      >
-        <div class="i-codicon-chrome-minimize" />
-      </span>
-      <span
-        title={isWindowMaximized ? 'Restore' : 'Maximize'}
-        class="hover:bg-hoverOverlay active:bg-hoverOverlayDarker dark:hover:bg-darkHoverOverlay dark:active:bg-darkHoverOverlayDarker"
-        on:click={toggleMaximize}
-      >
-        {#if isWindowMaximized}
-          <div class="i-codicon-chrome-restore" />
-        {:else}
-          <div class="i-codicon-chrome-maximize" />
-        {/if}
-      </span>
-    </span>
-  </div>
-{/if}
-
 <!-- Sidebar toggle, only visible on small screens -->
 <div
   id="sidebarToggle"
@@ -283,20 +198,18 @@
       src="tauri_logo.png"
       alt="Tauri logo"
     />
-    {#if !isWindows}
-      <a href="##" class="nv justify-between h-8" on:click={toggleDark}>
-        {#if isDark}
-          Switch to Light mode
-          <div class="i-ph-sun" />
-        {:else}
-          Switch to Dark mode
-          <div class="i-ph-moon" />
-        {/if}
-      </a>
-      <br />
-      <div class="bg-white/5 h-2px" />
-      <br />
-    {/if}
+    <a href="##" class="nv justify-between h-8" on:click={toggleDark}>
+      {#if isDark}
+        Switch to Light mode
+        <div class="i-ph-sun" />
+      {:else}
+        Switch to Dark mode
+        <div class="i-ph-moon" />
+      {/if}
+    </a>
+    <br />
+    <div class="bg-white/5 h-2px" />
+    <br />
 
     <a
       class="nv justify-between h-8"
