+++ tooling/cli/schema.json
@@ -2508,11 +2508,6 @@
           "items": {
             "type": "string"
           }
-        },
-        "enableTauriAPI": {
-          "description": "Enables access to the Tauri API.",
-          "default": false,
-          "type": "boolean"
         }
       },
       "additionalProperties": false
